Item List management
This small application allow you to create, load, save or email, list of
items like shopping products, things to do and so on.
It could be used as scrum while every person in a daily meeting will
load his weekly tasks and edit, remove or add some new tasks.
This is the beta version and it suppose or going to contain more
features in the future.